Sway/requirements.txt
google-generativeai>=0.8.0
requests>=2.31.0
tomllib
pathlib